orchestrator/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
markdown==3.3.4
django-filter==21.1
psycopg2-binary==2.9.1
gunicorn==20.1.0
kubernetes==21.7.0
django-typed-models==0.11.1
django-apscheduler==0.6.0
apscheduler==3.8.1
kafka-python==2.0.2
django-cleanup==5.2.0
